Add "Required Permissions To Load" to Fenced Frame Config.
Update FencedFrameConfig to include a new field:
required_permissions_to_load_. Previously, opaque-ads fenced frames
were always loaded with a static list of permissions which were always
enabled regardless of whether or not they were necessary. This was done
so that fenced frames loaded with an API can have access to the
necessary reporting tools (such as private aggregation).

However, not all APIs require all features to be enabled in order to
function properly. For instance, shared storage is one of the enabled
features, but FLEDGE does not need it.

With this change, APIs can now set the specific permissions they
require for a fenced frame to be enabled, and include them directly in
the configuration. This ensures that only the necessary permissions are
enabled, which enhances security and performance while reducing
unnecessary overhead.

Fenced frames created without an API will now not allow any features in
their permissions policies.

This CL updates various tests to respond to this change.

Change-Id: Ifb3cadc561d1de178ebbc540073f27b9b026ccac