[GridNG] Handle opposite writing direction subgrids
We were not previously accounting for opposite direction writing modes
in subgrids. This change reverses the order of set/range lookup and grid
geometry, and adjusts margin-border-padding in grid geometry to account
for opposing directions.

There will be followup work necessary for mapping orthogonal subgrids
while also accounting for writing direction. For now, there's a
condition blocking this from going down the new path as it caused a
regression.

A new test was added specifically for this behavior, as existing
tests only checked for this indirectly.

Bug: 618969
Change-Id: I3c3c8e99268beb5cfee81cb36ff810ec6c496cdb