[no-vary-search-hint] Add WPT test for No-Vary-Search hint functionality
In order to add a WPT test we have to create a situation where the
navigation to a potentially matching URL happens before the browser
receives the headers, but after the browser makes the prefetch request.

In order to synchronize the navigation we are introducing a new
python handler which will synchronize the in flight prefetch with the
WPT test by using request.server.stash.

Flow:
* prefetch prefetch_nvs_hint.py?uuid=...&nvs_header=...&otherqueryparams
* the prefetch request above includes no_vary_search_hint in the speculation
  rules
* the server blocks progress on this prefetch request on the server side so
  from the browser perspective the server is "thinking"
* the test starts navigation to
  prefetch_nvs_hint.py?uuid=...&nvs_header=...&otherdifferentqueryparams.
  This navigation matches by No-Vary-Search hint the above in
  progress prefetch.
* the test fetches prefetch_nvs_hint.py?uuid=...&unblock="unblock"
  which unblocks the in progress prefetch so that the in-progress
  navigation can continue

Bug: 1416659
Change-Id: I1b33b2bb8c649744f59ff6fda9f45cabeaa88847